requirements.txt
# =============== 核心依赖 ===============
langchain==0.1.14
langchain-core==0.1.33
langchain-community==0.0.34
langchain-openai==0.1.4
langchain-huggingface==0.0.42

# =============== 模型和嵌入 ===============
openai==1.3.8
sentence-transformers==2.2.2
transformers==4.35.2

# =============== 向量数据库 ===============
faiss-cpu==1.7.4
numpy==1.24.3

# =============== 文档处理 ===============
pypdf==4.0.1

# =============== 联网搜索 ===============
tavily-python==0.3.10

# =============== 基础库 ===============
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
huggingface-hub==0.19.4